+++ tests/ExceptionHandlerTest.php
@@ -3,11 +3,15 @@
 /**
  * @author Jared King <j@jaredtking.com>
  *
- * @link http://jaredtking.com
+ * @see http://jaredtking.com
  *
  * @copyright 2015 Jared King
  * @license MIT
  */
+
+namespace Infuse\Test;
+
+use Exception;
 use Infuse\Application;
 use Infuse\ExceptionHandler;
 use Infuse\Request;