+++ tests/HasAppTest.php
@@ -3,13 +3,17 @@
 /**
  * @author Jared King <j@jaredtking.com>
  *
- * @link http://jaredtking.com
+ * @see http://jaredtking.com
  *
  * @copyright 2015 Jared King
  * @license MIT
  */
+
+namespace Infuse\Test;
+
 use Infuse\Application;
 use Infuse\HasApp;
+use Mockery;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 
 class HasAppTest extends MockeryTestCase