+++ tests/MethodNotAllowedHandlerTest.php
@@ -3,11 +3,14 @@
 /**
  * @author Jared King <j@jaredtking.com>
  *
- * @link http://jaredtking.com
+ * @see http://jaredtking.com
  *
  * @copyright 2015 Jared King
  * @license MIT
  */
+
+namespace Infuse\Test;
+
 use Infuse\Application;
 use Infuse\MethodNotAllowedHandler;
 use Infuse\Request;