+++ tests/Services/PdoTest.php
@@ -3,14 +3,18 @@
 /**
  * @author Jared King <j@jaredtking.com>
  *
- * @link http://jaredtking.com
+ * @see http://jaredtking.com
  *
  * @copyright 2015 Jared King
  * @license MIT
  */
+
+namespace Infuse\Test\Services;
+
 use Infuse\Application;
 use Infuse\Services\Pdo as PdoService;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
+use PDO;
 
 class PdoTest extends MockeryTestCase
 {